NOTE/AUTO/requirements.txt
# 이미지 배경 투명화 프로그램 의존성 (모든 이미지 형식 지원)
Pillow>=9.0.0
numpy>=1.21.0
